Feat/v1 monitoring - Completed Initial Center and Outpost Monitoring Setup (#5)
### Work Complete

#### Ansible Role Monitor
Sub-Role: Center
This creates my central monitoring stack of Grafana, Loki, Mimir, Tailscale, Alloy which offer a central station for storing, viewing, and querying all my data. It also scrapes data from the server it is placed on which currently is the CLOUD server.

A socket-proxy with more restrained capability and no network connections was setup so that Alloy and Traefik would not have write capability to the docker socket.

Loki and Mimir are configured with multi-tenancy and data sources setup per tenant and service in Grafana. This way not only is incoming traffic only able to write to the allowed endpoint at /api/v1/push for each, the scope of what can be written to is regulated also based on the incoming X-Scope-OrgID header.

Sub-Role: Outpost
This creates my outpost monitoring stack of Traefik, Tailscale, Alloy, which scrapes data from the server it is placed on (every server other than the server CENTER is on)  and proxies it over Tailscale Traefik to reach to the monitor center in a regulated way with monitoring on the traffic.

#### Data Collected
Logs
- Docker stdout via Alloy - Loki Docker source

Metrics
- Docker container performance via Alloy - Cadvisor Source
- Traefik prometheus metrics via Alloy - Prometheus Scrape direct to Traefik instances
- Tailscale prometheus client metrics via Alloy - Prometheus Scrape direct to Tailscale instances

#### Ansible Role Edgeshark & Wireshark Setup on Dev Machine
This ansible role creates the Edgeshark stack on a server where is run giving me an overview of all the container network connection over the server and ability to connection to any specific interface on the server including notably here those in the docker containers that are part of specific connections. My dev machine's nixos configs were updated for wireshark support with the Edgeshark extension outside of this repo to work with created Edgeshark stack(s) and Tailscale grants were setup for it two way between Edgeshark and my dev machine for the required two way direct connection.


#### Tailscale Direct Connections - Proxmox Interfaces Management & OPNsense Router Configuration & Tailscale Container Configuration
A file was made for manual management of vlan interfaces in proxmox to be used by my terraform plans for creating the interfaces needed for macvlan connections on my servers. Those macvlan networks are setup in the ansible tasks plus docker composes for Tailscale containers. The default route is updated in all Tailscale containers to the macvlan interface. The prerequisite vlan devices, interfaces, and dhcp configs are setup in OPNsense for each isolated Tailscale container.

#### Tailscale ACLs Configuration
- Containers with outbound capability have access to and use Tailscale DNS where the dnsmasq instances from the ansible role: network-center have been used.
- The two way Edgeshark connection was allowed.
- The monitor center alloy instance was allowed to scrape Tailscale and Traefik instance metrics on its server.
- The monitor outpost alloy instance was allowed to scrape Tailscale and Traefik instance metrics on its server.
- The monitor outpost tailscale traefik instance (sidecar setup) was allowed to reach the CLOUD server network center Traefik instance.

Closes: #3 
[V1: Monitoring Setup](https://github.com/Nero-Study-Hat/Homelab/issues/3)


### Work Not Done Here
- Traces Data Collection
- Profiles Data Collection
- SIEM setup
- Automated Alerts & Responses
- Custom Grafana Dashboards & Queries